Move update functions to the module
